tests/wpt/web-platform-tests/css/css-grid/abspos/positioned-grid-items-should-not-create-implicit-tracks-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Positioned grid items should not create implicit tracks</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#abspos" title="9. Absolute Positioning">
<meta name="assert" content="This test checks that positioned items shouldn't create implicit tracks on the grid.">
<link href="support/grid.css" rel="stylesheet">
<style>

.grid {
  grid-auto-columns: 100px;
  grid-auto-rows: 50px;
  width: 400px;
  height: 300px;
  /* Ensures that the grid container is the containing block of the absolutely positioned grid children. */
  position: relative;
}

.absolute {
  position: absolute;
}

.seventhRowFourthColumn {
  background-color: coral;
  grid-column: 4;
  grid-row: 7;
}

</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<body onload="checkLayout('.grid')">

<div id="log"></div>

<div class="grid">
  <div class="sizedToGridArea absolute seventhRowFourthColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="400" data-expected-height="300">
  </div>
  <div class="autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="50">
  </div>
  <div class="autoRowAutoColumn"
    data-offset-x="0" data-offset-y="50" data-expected-width="100" data-expected-height="50">
  </div>
  <div class="autoRowAutoColumn"
    data-offset-x="0" data-offset-y="100" data-expected-width="100" data-expected-height="50">
  </div>
  <div class="autoRowAutoColumn"
    data-offset-x="0" data-offset-y="150" data-expected-width="100" data-expected-height="50">
  </div>
</div>

</body>
